Server: add production settings
