+++ packages/server/.babelrc
@@ -1,3 +1,14 @@
 {
-  "presets": ["@babel/preset-env"]
+  "presets": [
+    [
+      "@babel/preset-env",
+      {
+        "targets": {
+          "node": true
+        },
+        "useBuiltIns": "entry",
+        "corejs": 3
+      }
+    ]
+  ]
 }
