+++ packages/server/package.json
@@ -9,20 +9,27 @@
   },
   "scripts": {
     "build": "babel ./src --out-dir ./build --source-maps --copy-files",
-    "start": "nodemon --ignore ./build --exec babel-node ./bin/www"
+    "dev": "nodemon --ignore ./build --exec babel-node ./src/index.js",
+    "start": "node ./build"
   },
   "dependencies": {
     "@aragon/court": "1.1.0",
     "@aragon/court-backend-shared": "^0.0.1",
+    "@promster/express": "^4.0.0",
+    "@promster/server": "^4.0.0",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
+    "core-js": "^3.6.2",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "helmet": "^3.21.2",
     "jsonwebtoken": "^8.5.1",
     "morgan": "^1.9.1",
     "pg": "^7.15.1",
     "pg-hstore": "^2.3.3",
+    "prom-client": "^11.5.3",
+    "regenerator-runtime": "^0.13.3",
     "sequelize": "^5.21.3",
     "web3-utils": "^1.2.4"
   },
